Migrate to modern python project tools
This removes requirements.txt in favor of pyproject.toml only, adds Ruff
and mypy as the linter/formatter and type checker as well as fixing
issues they raised, and removes the multiple packaged scripts in favor
of a single oresat-configs.

Having only one script entry point let the script subcommand command
line argument setup be simplified from a two step process to a single
function.

The type annotations could be improved, I used a lot of asserts to
convince the checker that it's doing the right thing but ideally there
should be less or none of them. This is partially because canopen's type
annotations are incomplete or wrong in some areas, and partially because
3.10's match will make many cases easier to handle.

Ruff had a lot of good suggestions and I applied most of the straight
forward ones here. Some of the more involved ones (pathlib, pytest) I've
left for later commits.