Switch from unittest to pytest
pytest is the current gold standard for testing in python. The immediate
benefits of switching that we can make use of here is using regular
asserts instead of unittest asserts and fixtures. Fixtures mean that we
can simplify testing across all oresat missions and cards.

While I was here as a result of the paths cleanup I've added tests for
most scripts. Some additional paths related refactoring was added here
as well. I've also added type checking to the tests directory.

As a result two bugs were found:
- In c3.yaml a field was called names when it should have been name
- In gen_fw_files it assumed pdos were contiguous when they can be
  sparse.