Add verbose flag for od print script
This extends oresat-configs od with --verbose which will print every
field of an object dictionary object in addition to just the indexes.
This also adds description outputs in all cases.

Using this I found that arrays and records weren't having their
descriptions added so I fixed that while I was here.