Refactor PDO generation
PDOs as implemented in the configs have a lot of syntactic sugar - each
one needs associated objects and each TPDO needs an RPDO and vice versa.
The previous implementation unsugared at the same time as creating the
OD entries and this lead to an implementation that I found difficult to
follow. This divides the process into two steps:
- Unsugar into CardConfig
- Translate CardConfig directly into an ObjectDictionary. This is much
  easier now that there are fewer special cases to consider.

This implementation has the added benefit of being 1/3 faster (.16s ->
.11s under profile). Also reading the yaml files were rearranged so each
one only gets read once so the caching of CardConfig is no longer
needed.