Formalize card name aliases
This unifies card name parsing in a single method instead of leaving it
up to each script to come up with their own implementation. This allows
us to have a standard set of aliases and easily use them wherever we
need to refer to a card by name.

Behavior for scripts is mostly the same as before except that where
applicable they accept a wider set of names. gen_fw_files can now
generate the OD for every card, not just the embedded ones. I see no
reason to apply that restriction in the script, we can continue using it
for only the cards we need to.

Since this isn't the main goal of this series of commits there are two
major enhancements I left for a later standalone PR:
- Moving the alias list from hardcoded in the method to the yaml
  configs. When the configs move to being owned by each project then
  they can have control over how they're referred to, and it will be
  easy to add new ones
- A script (or enhancement to the cards script) that lists all the
  aliases.